Optimized CvPlayer::updateInventionEffectCache()
If given a civic as argument, it will only recalculate array data affected by that civic.
Example:
If the civic returns 0 for all getAllowsYields(), then the array will not change.
If getAllowsYields() isn't always 0, it will loop though all yields and skip most of the calculations for IDs, where getAllowsYields() is 0.

TODO: make CvCivic use JIT to store the data in question to gain a more reliable test for all 0 arrays.